feat(2.0): architecture redesign Ref: #108
