+++ package.json
@@ -66,8 +66,8 @@
     "langchain": ">=0.0.75"
   },
   "devDependencies": {
-    "@commitlint/cli": "^17.5.1",
-    "@commitlint/config-conventional": "^17.4.4",
+    "@commitlint/cli": "^17.7.1",
+    "@commitlint/config-conventional": "^17.7.0",
     "@types/lodash": "^4.14.194",
     "@types/node": "^18.15.5",
     "@types/node-fetch": "^2.6.4",
@@ -88,12 +88,13 @@
     "prettier": "^2.8.7",
     "ts-node": "^10.9.1",
     "tsup": "^6.7.0",
-    "typescript": "^5.0.4",
-    "vitest": "^0.31.1"
+    "typescript": "latest",
+    "vitest": "^0.34.2"
   },
   "dependencies": {
     "axios": "^1.3.4",
     "axios-cache-interceptor": "^1.0.1",
+    "callable-instance": "^2.0.0",
     "cross-fetch": "^3.1.6",
     "form-data": "^4.0.0",
     "promise-retry": "^2.0.1",
