+++ src/services/BaseService.ts
@@ -1,0 +1,9 @@
+import { FetchService } from './FetchService.js';
+
+export abstract class BaseService {
+  protected readonly fetcher: FetchService;
+
+  constructor(fetcher: FetchService) {
+    this.fetcher = fetcher;
+  }
+}
