+++ src/tests/integration/client.test.ts
@@ -22,15 +22,17 @@ describe('client', () => {
   let client: Client;
   beforeEach(() => {
     client = new Client({
-      endpoint: MOCK_ENDPOINT,
-      apiKey: 'foobar',
+      config: {
+        endpoint: MOCK_ENDPOINT,
+        apiKey: 'foobar',
+      },
     });
   });
 
   describe('generate', () => {
     describe('config', () => {
       test('should read the config', async () => {
-        const config = await client.generateConfig();
+        const config = await client.generate.config.get();
         expect(config).toMatchObject({ model_id: 'foobar' });
       });
 
@@ -42,9 +44,7 @@ describe('client', () => {
             random_seed: 8,
           },
         };
-        const config = await client.generateConfig(input, {
-          strategy: 'replace',
-        });
+        const config = await client.generate.config.replace(input);
         expect(config).toMatchObject(input);
       });
 
@@ -55,9 +55,7 @@ describe('client', () => {
             random_seed: 8,
           },
         };
-        const config = await client.generateConfig(input, {
-          strategy: 'merge',
-        });
+        const config = await client.generate.config.merge(input);
         expect(config).toMatchObject({ model_id: 'foobar', ...input });
       });
 
@@ -69,18 +67,16 @@ describe('client', () => {
             random_seed: 8,
           },
         };
-        const replacedConfig = await client.generateConfig(input, {
-          strategy: 'replace',
-        });
+        const replacedConfig = await client.generate.config.replace(input);
         expect(replacedConfig).toMatchObject(input);
 
-        const config = await client.generateConfig({ reset: true });
+        const config = await client.generate.config.reset();
         expect(config).toMatchObject({ model_id: 'foobar' });
       });
     });
 
     test('should return single output for a single input', async () => {
-      const data = await client.generate({
+      const data = await client.generate.generate({
         model_id: 'bigscience/bloom',
         input: 'Hello, World',
       });
@@ -100,7 +96,7 @@ describe('client', () => {
         },
       ];
 
-      const outputs = await Promise.all(client.generate(inputs));
+      const outputs = await Promise.all(client.generate.generate(inputs));
       expect(outputs.length).toBe(inputs.length);
       outputs.forEach((output) => {
         expect(output).toMatchObject(generateStore);
@@ -111,7 +107,7 @@ describe('client', () => {
   describe('tokenize', () => {
     test('should return tokenize info', () => {
       expect(
-        client.tokenize({
+        client.tokenizer.tokenize({
           input: 'Hello, how are you? Are you okay?',
           model_id: 'google/flan-t5-xl',
         }),
@@ -121,13 +117,13 @@ describe('client', () => {
 
   describe('models', () => {
     test('should return some models', async () => {
-      const models = await client.models();
+      const models = await client.models.list();
       expect(models.length).not.toBeEmpty();
     });
 
     test('should return details for a given model', async () => {
       const id = modelsStore[0].id;
-      const details = await client.model({ id });
+      const details = await client.models.get({ id });
       expect(details).toHaveProperty('id', id);
     });
   });
@@ -137,7 +133,7 @@ describe('client', () => {
       new Promise((done) => {
         expect.assertions(tunesStore.length * 2);
         let count = tunesStore.length;
-        client.tunes((err, tune) => {
+        client.tunes.list((err, tune) => {
           expect(err).toBeFalsy();
           expect(tune).toBeDefined();
           if (--count === 0) done(undefined);
@@ -146,7 +142,7 @@ describe('client', () => {
 
     test('should list all tunes', async () => {
       const allTunes = [];
-      for await (const tune of client.tunes()) {
+      for await (const tune of client.tunes.list()) {
         allTunes.push(tune);
       }
       expect(allTunes.length).toBe(tunesStore.length);
@@ -156,13 +152,13 @@ describe('client', () => {
     });
 
     test('should list all tune methods', async () => {
-      const tuneMethods = await client.tuneMethods();
+      const tuneMethods = await client.tunes.listMethods();
       expect(tuneMethods.length).toBe(tuneMethodsStore.length);
     });
 
     test('should show details of a tune', async () => {
       const { id } = tunesStore[0];
-      const tune = await client.tune({ id });
+      const tune = await client.tunes.get({ id });
       expect(tune).toHaveProperty('id', id);
     });
 
@@ -171,7 +167,7 @@ describe('client', () => {
         id,
         assets: { encoder },
       } = tunesStore[0];
-      const tune = await client.tune({ id });
+      const tune = await client.tunes.get({ id });
       expect.assertions(1);
       if (tune.status === 'COMPLETED') {
         const content = await tune.downloadAsset('encoder');
@@ -183,21 +179,21 @@ describe('client', () => {
     });
 
     test('should create a tune', async () => {
-      const tune = await client.tune(dummyTune);
+      const tune = await client.tunes.create(dummyTune);
       expect(tunesStore.map(({ id }) => id)).toContainEqual(tune.id);
     });
 
     test('should delete a tune', async () => {
       const { id } = tunesStore[1];
-      await client.tune({ id }, { delete: true });
+      await client.tunes.remove({ id });
       expect(tunesStore.map(({ id }) => id)).not.toContain({ id });
     });
 
     test('should not get a deleted tune', async () => {
-      const { id } = await client.tune(dummyTune);
-      await client.model({ id });
-      await client.tune({ id }, { delete: true });
-      expect(client.tune({ id })).toReject();
+      const { id } = await client.tunes.create(dummyTune);
+      await client.models.get({ id });
+      await client.tunes.remove({ id });
+      expect(client.tunes.get({ id })).toReject();
     });
   });
 
@@ -206,7 +202,7 @@ describe('client', () => {
       new Promise((done) => {
         expect.assertions(promptTemplatesStore.length * 2);
         let count = promptTemplatesStore.length;
-        client.promptTemplates((err, template) => {
+        client.promptTemplates.list((err, template) => {
           expect(err).toBeFalsy();
           expect(template).toBeDefined();
           if (--count === 0) done(undefined);
@@ -215,7 +211,7 @@ describe('client', () => {
 
     test('should list all prompt templates using for-await loop', async () => {
       const templates = [];
-      for await (const promptTemplate of client.promptTemplates()) {
+      for await (const promptTemplate of client.promptTemplates.list()) {
         templates.push(promptTemplate);
       }
       expect(templates.length).toBe(promptTemplatesStore.length);
@@ -233,12 +229,12 @@ describe('client', () => {
 
     test('should show details of a prompt template', async () => {
       const { id } = promptTemplatesStore[0];
-      const promptTemplate = await client.promptTemplate({ id });
+      const promptTemplate = await client.promptTemplates.get({ id });
       expect(promptTemplate).toHaveProperty('id', id);
     });
 
     test('should create a prompt template', async () => {
-      const template = await client.promptTemplate({
+      const template = await client.promptTemplates.create({
         name: 'my template',
         value: 'Hello {{name}}!',
       });
@@ -248,7 +244,7 @@ describe('client', () => {
     });
 
     test('should update the prompt template', async () => {
-      const template = await client.promptTemplate({
+      const template = await client.promptTemplates.update({
         id: promptTemplatesStore[1].id,
         value: promptTemplatesStore[1].value,
         name: 'Greeting template!',
@@ -258,12 +254,12 @@ describe('client', () => {
 
     test('should delete a prompt template', async () => {
       const { id } = promptTemplatesStore[1];
-      await client.promptTemplate({ id }, { delete: true });
+      await client.promptTemplates.remove({ id });
       expect(promptTemplatesStore.map(({ id }) => id)).not.toContain({ id });
     });
 
     test('should execute a prompt template', async () => {
-      const results = await client.promptTemplateExecute({
+      const results = await client.promptTemplates.print({
         inputs: ['1+1', '2+2', '3+3'],
         template: {
           value: 'Hello {{name}}, how much is {{input}}?',
@@ -282,31 +278,31 @@ describe('client', () => {
     });
 
     test('should not get a deleted prompt template', async () => {
-      const { id } = await client.promptTemplate({
+      const { id } = await client.promptTemplates.create({
         name: 'my template',
         value: 'Hello {{name}}!',
       });
-      await client.promptTemplate({ id });
-      await client.promptTemplate({ id }, { delete: true });
-      expect(client.tune({ id })).toReject();
+      await client.promptTemplates.get({ id });
+      await client.promptTemplates.remove({ id });
+      expect(client.tunes.get({ id })).toReject();
     });
 
     test('should get an updated prompt template', async () => {
       const oldName = 'old name';
-      const { id } = await client.promptTemplate({
+      const { id } = await client.promptTemplates.create({
         name: oldName,
         value: 'Hello {{name}}!',
       });
-      const oldTemplate = await client.promptTemplate({ id });
+      const oldTemplate = await client.promptTemplates.get({ id });
       expect(oldTemplate).toHaveProperty('name', oldName);
 
       const newName = 'old name';
-      await client.promptTemplate({
+      await client.promptTemplates.update({
         id,
         name: newName,
         value: 'Hello {{name}}!',
       });
-      const newTemplate = await client.promptTemplate({ id });
+      const newTemplate = await client.promptTemplates.get({ id });
       expect(newTemplate).toHaveProperty('name', newName);
     });
   });
@@ -316,7 +312,7 @@ describe('client', () => {
       new Promise((done) => {
         expect.assertions(historyStore.length * 2);
         let count = historyStore.length;
-        client.history((err, entry) => {
+        client.history.list((err, entry) => {
           expect(err).toBeFalsy();
           expect(entry).toBeDefined();
           if (--count === 0) done(undefined);
@@ -325,7 +321,7 @@ describe('client', () => {
 
     test('should list all requests from the past using for-await loop', async () => {
       const entries = [];
-      for await (const entry of client.history()) {
+      for await (const entry of client.history.list()) {
         entries.push(entry);
         expect(entry).toBeDefined();
       }
@@ -335,25 +331,25 @@ describe('client', () => {
 
   describe('cross method', () => {
     test('should not get deleted tune via model', async () => {
-      const { id } = await client.tune(dummyTune);
-      await client.model({ id });
-      await client.tune({ id }, { delete: true });
-      expect(client.model({ id })).toReject();
+      const { id } = await client.tunes.create(dummyTune);
+      await client.models.get({ id });
+      await client.tunes.remove({ id });
+      expect(client.models.get({ id })).toReject();
     });
 
     test('should not get deleted tune in models list', async () => {
-      const { id } = await client.tune(dummyTune);
-      const modelsWithTune = await client.models();
+      const { id } = await client.tunes.create(dummyTune);
+      const modelsWithTune = await client.models.list();
       expect(modelsWithTune.map(({ id }) => id)).toContain(id);
-      await client.tune({ id }, { delete: true });
-      const models = await client.models();
+      await client.tunes.remove({ id });
+      const models = await client.models.list();
       expect(models.map(({ id }) => id)).not.toContain(id);
     });
 
     test('should get newly added tune in models list', async () => {
-      await client.models();
-      const { id } = await client.tune(dummyTune);
-      const models = await client.models();
+      await client.models.list();
+      const { id } = await client.tunes.create(dummyTune);
+      const models = await client.models.list();
       expect(models.map(({ id }) => id)).toContain(id);
     });
   });
